[feat] Update user_manager module
- 회원가입 로그인 기능을 user_manager 모듈로 분리
